move components to proper folder
